Fix include order of boost python.
Due to the fact that the function mrt_add_python_api
searches for BoostPython and PythonLibs and adds those
to the include directories they are listed before the
project include folder. This leads to problems, if there is
a header in python which has the same name as in the project.
This patch resolves that issue by removing the find_package
handling and moving it auto depend.