Improved Qt experience.
This PR makes it possbile to use the standard mrt cmake file for QT
projects. To achive that, the find script uses the targets insteads of
the libraries for QT_LIBRARIES because the targets also adds
compiler flags, etc. In addition, the source directory of executables
are added so that ui files referencing header files in the source
tree are working out of the box.