Improved CUDA support.
Beginning with cmake 3.9 the native CUDA language support in cmake is used to
build cuda files. This is achived by building a separate shared library for
the cuda files as this is the only way to select differnt GCC version, as the
current CUDA (10.2) version is not compatible with gcc 8.

The problem which arises which libraries needs to be linked into the cuda
shared library. Linking just all libraries does not work as some external
libraries like PCL carries flags in target_link_libraries which nvcc does
not understand. Using no library (like it is in the current approach) work
in many cases but not when e.g. OpenCV is used in CUDA code.

This MR adds a new flag to the <external> section in the package.xml
<cuda_depend> where one can specify the librarys which one would like
to link to CUDA code.

Also the behaviour when cuda code in executables is simplified (only
building a single shared library instead of an object file linked to a
shared library file).