Fix build for non-x86 architectures
When MRT_ARCH is not explicitly set, UseMrtStdCompilerFlags.cmake
currently forces the target architecture to be tuned to the Sandy Bridge
(x86) architecture.

This patch constraints this behaviour to x86-based targets. This allow
projects based on MRT Cmake Modules to be rebuild transparently in other
architectures without any extra build settings.

Note that this change works with both cross-compilation and native
non-x86 builds.

See #7.

Change-Id: Ibc1ed1eaf1c014df4c44d774ea8fd3760a1679e7
Signed-off-by: Filipe Rinaldi <filipe.rinaldi@arm.com>