Create an own testing suite for mrt_cmake_modules
The suite can handle ctest (builds the source automatically).
Rostest, gtest and nosetest are supported similar to catkin (outputting xml results in the same dir).
Additionally, coverage results are generated and output to the command line, together with an url that can be opened in the browser. This works for both ctest and normal (catkin) builds.
If python-coverage is installed, python coverage is included as well.
Building with the toplevel cmakelists works as well.